python/requirements.txt
blib-py
netCDF4
numpy
radar-data
tqdm
wheel
